requirements.txt
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.5.2
astroid==2.6.2
async-generator==1.10
attrs==21.2.0
Automat==20.2.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
curlify2==1.0.1
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.3
exceptiongroup==1.0.0
fastapi==0.86.0
filelock==3.8.0
Flask==2.2.2
h11==0.14.0
httptools==0.5.0
humanize==4.0.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==5.0.0
incremental==22.10.0
iniconfig==1.1.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.2
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keyboard==0.13.5
lazy-object-proxy==1.6.0
lxml==4.9.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
orjson==3.8.1
packaging==21.0
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.10.2
PyDispatcher==2.0.6
PyGithub==1.55
Pygments==2.9.0
PyJWT==2.5.0
pylint==2.9.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==7.2.0
python-dateutil==2.8.1
python-decouple==3.6
python-dotenv==0.21.0
python-gitlab==3.3.0
python-multipart==0.0.5
pywin32==301
pywinpty==1.1.3
PyYAML==6.0
pyzmq==22.1.0
queuelib==1.6.2
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.10.0
responses==0.12.1
review==1510
reviews==0.4.2
rich==12.1.0
scanapi==2.8.0
Scrapy==2.7.0
Send2Trash==1.7.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starlette==0.20.4
terminado==0.10.1
testpath==0.5.0
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.0.5
Twisted==22.8.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2022.2
ujson==5.5.0
urllib3==1.26.6
uvicorn==0.18.3
w3lib==2.0.1
watchfiles==0.18.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.10.0
zope.interface==5.5.0
